logsy_server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
asyncpg
aiohttp
requests
grpcio
grpcio-tools